Move `retry` to util module
Again, avoid single-function modules.
