+++ cognite/extractorutils/uploader/files.py
@@ -23,7 +23,6 @@ from cognite.client.data_classes import Event, FileMetadata
 from cognite.client.exceptions import CogniteAPIError
 from requests import ConnectionError
 
-from cognite.extractorutils.retry import retry
 from cognite.extractorutils.uploader._base import (
     RETRIES,
     RETRY_BACKOFF_FACTOR,
@@ -41,6 +40,7 @@ from cognite.extractorutils.uploader._metrics import (
     FILES_UPLOADER_QUEUED,
     FILES_UPLOADER_WRITTEN,
 )
+from cognite.extractorutils.util import retry
 
 
 class FileUploadQueue(AbstractUploadQueue):
